Create umbrella app and mae data model to child app
