+++ .gitignore
@@ -1,17 +1,17 @@
 # The directory Mix will write compiled artifacts to.
-/_build/
+**/_build/
 
 # If you run "mix test --cover", coverage assets end up here.
-/cover/
+**/cover/
 
 # The directory Mix downloads your dependencies sources to.
-/deps/
+**/deps/
 
 # Where third-party dependencies like ExDoc output generated docs.
-/doc/
+**/doc/
 
 # Ignore .fetch files in case you like to edit your project deps locally.
-/.fetch
+**/.fetch
 
 # If the VM crashes, it generates a dump, let's ignore it too.
 erl_crash.dump
@@ -22,4 +22,4 @@ erl_crash.dump
 # Ignore package tarball (built via "mix hex.build").
 next_next_gen-*.tar
 
-/.elixir_ls/
+**/.elixir_ls/
