+++ apps/next_next_gen/mix.exs
@@ -1,0 +1,38 @@
+defmodule NextNextGen.MixProject do
+  use Mix.Project
+
+  def project do
+    [
+      app: :next_next_gen,
+      version: "0.1.0",
+      elixir: "~> 1.8",
+      start_permanent: Mix.env() == :prod,
+      deps: deps(),
+      aliases: aliases()
+    ]
+  end
+
+  # Run "mix help compile.app" to learn about applications.
+  def application do
+    [
+      extra_applications: [:logger],
+      mod: {NextNextGen.Application, []}
+    ]
+  end
+
+  # Run "mix help deps" to learn about dependencies.
+  defp deps do
+    [
+      {:postgrex, ">= 0.0.0"},
+      {:ecto_sql, "~> 3.0"},
+      {:jason, "~> 1.1"}
+    ]
+  end
+
+  defp aliases do
+    [
+      "ecto.setup": ["ecto.create", "ecto.migrate", "run priv/repo/seeds.exs"],
+      "ecto.reset": ["ecto.drop", "ecto.setup"]
+    ]
+  end
+end
